--- a/libcxx/test/std/containers/sequences/vector.bool/copy.pass.cpp
+++ b/libcxx/test/std/containers/sequences/vector.bool/copy.pass.cpp
@@ -11,12 +11,13 @@
 
 // vector(const vector& v);
 
-#include <vector>
+#include <array>
 #include <cassert>
+#include <vector>
 
-#include "test_macros.h"
-#include "test_allocator.h"
 #include "min_allocator.h"
+#include "test_allocator.h"
+#include "test_macros.h"
 
 template <class C>
 TEST_CONSTEXPR_CXX20 void test(const C& x) {
@@ -25,39 +26,56 @@ TEST_CONSTEXPR_CXX20 void test(const C& x) {
   LIBCPP_ASSERT(c.__invariants());
   assert(c.size() == s);
   assert(c == x);
+#if TEST_STD_VER >= 11
+  assert(c.get_allocator() ==
+         std::allocator_traits<typename C::allocator_type>::select_on_container_copy_construction(x.get_allocator()));
+#endif
 }
 
 TEST_CONSTEXPR_CXX20 bool tests() {
-  {
-    bool a[] = {0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0};
-    bool* an = a + sizeof(a) / sizeof(a[0]);
-    test(std::vector<bool>(a, an));
+  std::array<int, 5> a1   = {1, 0, 1, 0, 1};
+  std::array<int, 18> a2  = {0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0};
+  std::array<int, 33> a3  = {0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0};
+  std::array<int, 65> a4  = {0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0};
+  std::array<int, 299> a5 = {};
+  for (unsigned i = 0; i < a5.size(); i += 2)
+    a5[i] = 1;
+
+  // Tests for vector<bool> copy constructor with word size up to 5 (i.e., bit size > 256 on a 64-bit system)
+  { // Test with default std::allocator
+    test(std::vector<bool>(a1.begin(), a1.end()));
+    test(std::vector<bool>(a2.begin(), a2.end()));
+    test(std::vector<bool>(a3.begin(), a3.end()));
+    test(std::vector<bool>(a4.begin(), a4.end()));
+    test(std::vector<bool>(a5.begin(), a5.end()));
   }
-  {
-    std::vector<bool, test_allocator<bool> > v(3, true, test_allocator<bool>(5));
-    std::vector<bool, test_allocator<bool> > v2 = v;
-    assert(v2 == v);
-    assert(v2.get_allocator() == v.get_allocator());
+  { // Test with test_allocator
+    using A = test_allocator<bool>;
+    using C = std::vector<bool, A>;
+    test(C(a1.begin(), a1.end()));
+    test(C(a2.begin(), a2.end()));
+    test(C(a3.begin(), a3.end()));
+    test(C(a4.begin(), a4.end()));
+    test(C(a5.begin(), a5.end()));
   }
-#if TEST_STD_VER >= 11
-  {
-    std::vector<bool, other_allocator<bool> > v(3, true, other_allocator<bool>(5));
-    std::vector<bool, other_allocator<bool> > v2 = v;
-    assert(v2 == v);
-    assert(v2.get_allocator() == other_allocator<bool>(-2));
+  { // Test with other_allocator
+    using A = other_allocator<bool>;
+    using C = std::vector<bool, A>;
+    test(C(a1.begin(), a1.end()));
+    test(C(a2.begin(), a2.end()));
+    test(C(a3.begin(), a3.end()));
+    test(C(a4.begin(), a4.end()));
+    test(C(a5.begin(), a5.end()));
   }
-  {
-    bool a[] = {0, 1, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0};
-    bool* an = a + sizeof(a) / sizeof(a[0]);
-    test(std::vector<bool, min_allocator<bool>>(a, an));
+  { // Test with min_allocator
+    using A = min_allocator<bool>;
+    using C = std::vector<bool, A>;
+    test(C(a1.begin(), a1.end()));
+    test(C(a2.begin(), a2.end()));
+    test(C(a3.begin(), a3.end()));
+    test(C(a4.begin(), a4.end()));
+    test(C(a5.begin(), a5.end()));
   }
-  {
-    std::vector<bool, min_allocator<bool> > v(3, true, min_allocator<bool>());
-    std::vector<bool, min_allocator<bool> > v2 = v;
-    assert(v2 == v);
-    assert(v2.get_allocator() == v.get_allocator());
-  }
-#endif
 
   return true;
 }