--- a/libcxx/test/std/containers/sequences/vector.bool/move.pass.cpp
+++ b/libcxx/test/std/containers/sequences/vector.bool/move.pass.cpp
@@ -9,54 +9,70 @@
 // UNSUPPORTED: c++03
 
 // <vector>
+// vector<bool>
 
 // vector(vector&& c);
 
-#include <vector>
+#include <array>
 #include <cassert>
-#include "test_macros.h"
-#include "test_allocator.h"
+#include <vector>
+
 #include "min_allocator.h"
+#include "test_allocator.h"
+#include "test_macros.h"
+
+template <unsigned N, class A>
+TEST_CONSTEXPR_CXX20 void test(const A& a) {
+  std::vector<bool, A> v(N, false, a);
+  std::vector<bool, A> v0(N, false, a);
+  for (unsigned i = 1; i < N; i += 2) {
+    v[i]  = true;
+    v0[i] = true;
+  }
+  std::vector<bool, A> v2 = std::move(v);
+  assert(v2 == v0);
+  assert(v.empty()); // The moved-from vector is guarantted to be empty after move-construction
+  assert(v2.get_allocator() == v0.get_allocator());
+}
 
 TEST_CONSTEXPR_CXX20 bool tests() {
   test_allocator_statistics alloc_stats;
+
+  // Tests for move constructor with word size up to 5 (i.e., bit size > 256 for 64-bit system)
   {
-    std::vector<bool, test_allocator<bool> > l(test_allocator<bool>(5, &alloc_stats));
-    std::vector<bool, test_allocator<bool> > lo(test_allocator<bool>(5, &alloc_stats));
-    for (int i = 1; i <= 3; ++i) {
-      l.push_back(true);
-      lo.push_back(true);
-    }
-    std::vector<bool, test_allocator<bool> > l2 = std::move(l);
-    assert(l2 == lo);
-    assert(l.empty());
-    assert(l2.get_allocator() == lo.get_allocator());
+    using A = std::allocator<bool>;
+    test<5>(A());
+    test<18>(A());
+    test<33>(A());
+    test<65>(A());
+    test<299>(A());
   }
   {
-    std::vector<bool, other_allocator<bool> > l(other_allocator<bool>(5));
-    std::vector<bool, other_allocator<bool> > lo(other_allocator<bool>(5));
-    for (int i = 1; i <= 3; ++i) {
-      l.push_back(true);
-      lo.push_back(true);
-    }
-    std::vector<bool, other_allocator<bool> > l2 = std::move(l);
-    assert(l2 == lo);
-    assert(l.empty());
-    assert(l2.get_allocator() == lo.get_allocator());
+    using A = other_allocator<bool>;
+    test<5>(A(5));
+    test<18>(A(5));
+    test<33>(A(5));
+    test<65>(A(5));
+    test<299>(A(5));
   }
   {
-    std::vector<bool, min_allocator<bool> > l(min_allocator<bool>{});
-    std::vector<bool, min_allocator<bool> > lo(min_allocator<bool>{});
-    for (int i = 1; i <= 3; ++i) {
-      l.push_back(true);
-      lo.push_back(true);
-    }
-    std::vector<bool, min_allocator<bool> > l2 = std::move(l);
-    assert(l2 == lo);
-    assert(l.empty());
-    assert(l2.get_allocator() == lo.get_allocator());
+    using A = min_allocator<bool>;
+    test<5>(A());
+    test<18>(A());
+    test<33>(A());
+    test<65>(A());
+    test<299>(A());
   }
   {
+    using A = test_allocator<bool>;
+    test<5>(A(5, &alloc_stats));
+    test<18>(A(5, &alloc_stats));
+    test<33>(A(5, &alloc_stats));
+    test<65>(A(5, &alloc_stats));
+    test<299>(A(5, &alloc_stats));
+  }
+
+  { // Tests to verify the allocator statistics after move
     alloc_stats.clear();
     using Vect   = std::vector<bool, test_allocator<bool> >;
     using AllocT = Vect::allocator_type;
@@ -83,7 +99,6 @@ TEST_CONSTEXPR_CXX20 bool tests() {
       const AllocT& a2 = v2.get_allocator();
       assert(a2.get_id() == 101);
       assert(a2.get_data() == 42);
-
       assert(a1 == a2);
     }
   }