--- a/libcxx/test/std/containers/sequences/vector.bool/move.pass.cpp
+++ b/libcxx/test/std/containers/sequences/vector.bool/move.pass.cpp
@@ -24,15 +24,15 @@
 template <unsigned N, class A>
 TEST_CONSTEXPR_CXX20 void test(const A& a) {
   std::vector<bool, A> v(N, false, a);
-  std::vector<bool, A> v0(N, false, a);
+  std::vector<bool, A> original(N, false, a);
   for (unsigned i = 1; i < N; i += 2) {
-    v[i]  = true;
-    v0[i] = true;
+    v[i]        = true;
+    original[i] = true;
   }
   std::vector<bool, A> v2 = std::move(v);
-  assert(v2 == v0);
-  assert(v.empty()); // The moved-from vector is guarantted to be empty after move-construction
-  assert(v2.get_allocator() == v0.get_allocator());
+  assert(v2 == original);
+  assert(v.empty()); // The moved-from vector is guaranteed to be empty after move-construction
+  assert(v2.get_allocator() == original.get_allocator());
 }
 
 TEST_CONSTEXPR_CXX20 bool tests() {