<!-- New Version --> 
The current tests for `vector<bool>` fail to adequately cover realistic scenarios, as they are limited to cases of up to 3 bytes, representing less than 1/2 of a word size on a 64-bit system. However, most `vector<bool>` operations rely on code paths triggered only when handling multiple storage words (8 bytes each). To address this gap, this PR rewrites the tests for copy and move constructors, as well as allocator-extended copy and move constructors, ensuring that previously untested code paths are now thoroughly validated.

Note: This PR now focuses on improving test coverage, as the performance enhancements have been addressed in separate patches that optimize algorithms `ranges::{copy, move}` for `__bit_iterator`s  (e.g., #121013, #121109). For completeness, I have retained the history of this patch to provide full context.



<!-- Old Version -->
### ~~General description~~ 
~~This PR is part of a series aimed at significantly improving the performance of `vector<bool>`. Each PR focuses on enhancing a specific subset of operations, ensuring they are self-contained and easy to review. The main idea for performance improvements involves using word-wise implementation along with bit manipulation techniques, rather than solely using bit-wise operations in the previous implementation, resulting in substantial performance gains.~~ 



~~### Current PR~~
~~This PR significantly improves the performance of various constructors in `vector<bool>` by at least **500x** under a variety of practical storage word sizes. The improvements for the above tests are:~~

~~- Copy constructor: **576x**~~
~~- Copy constructor with extended allocator: **546x**~~
~~- Move constructor with extended allocator: **568x**~~ 


~~### Tests~~

~~The existing tests for `vector<bool>` are largely insufficient to cover realistic scenarios. They are limited to cases of up to 1~2 bytes. However, most `vector<bool>` operations involve code paths that are executed only when multiple storage words, 8 bytes each, are involved. To fix the test coverage, this PR completely rewrites the constructor tests to ensure that previously untested code paths are now properly covered.~~

~~Furthermore, this PR conducts benchmark testing for the move-assignment operator, demonstrating at least 500x performance improvement.~~ 

~~#### Before:~~


~~-------------------------------------------------------------------------------------~~
~~Benchmark                                           Time             CPU   Iterations~~
~~-------------------------------------------------------------------------------------~~
~~BM_CopyConstruct/vector_bool/5140480          8596063 ns      8563419 ns           78~~
~~BM_MoveConstruct/vector_bool/5140480            0.474 ns        0.473 ns   1470604919~~
~~BM_CopyConstruct_Alloc/vector_bool/5140480    8182327 ns      8169876 ns           83~~
~~BM_MoveConstruct_Alloc/vector_bool/5140480    8544951 ns      8537676 ns           83~~


~~#### After:~~
~~-------------------------------------------------------------------------------------~~
~~Benchmark                                           Time             CPU   Iterations~~
~~-------------------------------------------------------------------------------------~~
~~BM_CopyConstruct/vector_bool/5140480            14899 ns        14943 ns        46010~~
~~BM_MoveConstruct/vector_bool/5140480            0.451 ns        0.453 ns   1571837345~~
~~BM_CopyConstruct_Alloc/vector_bool/5140480      14979 ns        15033 ns        47913~~
~~BM_MoveConstruct_Alloc/vector_bool/5140480      15043 ns        15101 ns        45906~~