@winner245 2024-12-21
Thank you for your feedback. If I understand correctly, you are suggesting that I work on the general `std::copy` algorithm (and possibly std::move and others) for the `__bit_iterator` overload. I am a bit confused. Isn't this overload already implemented in `libcxx/include/__bit_reference`?