@philnik777 2025-01-20
I'm personally not a huge fan of this, since it adds another code path you have to know about to understand what vector does, especially since this doesn't improve anything over calling `std::copy` (assuming the optimization is fixed).