@ldionne 2025-01-20
I suggest we then wait for the underlying regression fix to land so we can reevaluate this patch. At that point, this would become a pure NFC refactoring IIUC, and we can evaluate it on that basis.