@philnik777 2025-01-20
> I suggest we then wait for the underlying regression fix to land so we can reevaluate this patch. At that point, this would become a pure NFC refactoring IIUC, and we can evaluate it on that basis.

Yes, that should be the case. If it's not then this would probably get my approval as well.

(Also, how the hell does github work? I wrote the second comment before the first one, but GH deleted that for me and I didn't feel like re-writing it, so I wrote a shorter message)