@winner245 2025-01-20
> > I suggest we then wait for the underlying regression fix to land so we can reevaluate this patch. At that point, this would become a pure NFC refactoring IIUC, and we can evaluate it on that basis.
>
> Yes, that should be the case. If it's not then this would probably get my approval as well.

Thanks @philnik777. I think your concern is valid, and I am fine with that. In that case, we will need to land #121013 first. Then this one becomes NFC or I might need to improve the tests for the copy/move constructors as the existing ones only considered up to 3 bytes, which is less than the word size of 8 bytes.